requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
bcrypt
python-multipart
sqlalchemy
pymysql
alembic
pydantic
pydantic-settings
python-dotenv
email-validator
jinja2
httpx
redis
pyotp
pytest
pytest-asyncio
